# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateTransactionRequest:
    properties:
      amount:
        type: number
      category_id:
        type: string
      date:
        type: string
      note:
        maxLength: 500
        type: string
      platform_id:
        type: string
      pocket_from:
        type: string
      pocket_to:
        type: string
      ref:
        maxLength: 100
        type: string
      type:
        enum:
        - income
        - expense
        - transfer
        type: string
      user_platform_from:
        type: string
      user_platform_to:
        type: string
    required:
    - amount
    - date
    - type
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive financial management system with smart allocation engine,
    transaction tracking, and detailed reports
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Coin Backend API
  version: "1.0"
paths:
...
  /v1/transactions:
    get:
      consumes:
      - application/json
      description: Get a list of user transactions with pagination
      parameters:
      - description: 'Limit (default: 10, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: 'Skip (default: 0)'
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transactions retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List user transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction for the authenticated user
      parameters:
      - description: Transaction details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new transaction
      tags:
      - Transactions
  /v1/transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - Transactions
    get:
      consumes:
      - application/json
      description: Get a specific transaction by ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - Transactions
  /v1/transactions/pocket/{pocket_id}:
    get:
      consumes:
      - application/json
      description: Get a list of transactions for a specific pocket
      parameters:
      - description: Pocket ID
        in: path
        name: pocket_id
        required: true
        type: string
      - description: 'Limit (default: 10, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: 'Skip (default: 0)'
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transactions retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List pocket transactions
      tags:
      - Transactions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"